+++ src/Controller/AppController.php
@@ -103,7 +103,7 @@ class AppController extends BaseController {
 		
 		//Sets the paginate limit and the maximum paginate limit
 		//See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
-		$this->paginate['limit'] = $this->paginate['maxLimit'] = $this->request->isAdmin() ? config('backend.records') : config('default.records');
+		$this->paginate['limit'] = $this->paginate['maxLimit'] = $this->request->isAdmin() ? config('admin.records') : config('default.records');
 		
 		parent::beforeFilter($event);
 	}